requirements.txt
Cython==3.0.6
QtPy==2.4.1
SecretStorage==3.3.3
arrow==1.3.0
contextlib2==21.6.0
debugpy==1.8.0
docutils==0.20.1
fqdn==1.5.1
ipywidgets==8.1.1
isoduration==20.11.0
jaraco.classes==3.3.0
jeepney==0.8.0
jsonpointer==2.4
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-lsp==2.2.1
jupyterlab-widgets==3.0.9
keyring==24.3.0
libclang==16.0.6
nh3==0.2.14
notebook==7.0.6
pkginfo==1.9.6
qtconsole==5.5.1
readme-renderer==42.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
tensorflow-io-gcs-filesystem==0.34.0
terminado==0.18.0
twine==4.0.2
types-python-dateutil==2.8.19.14
uri-template==1.3.0
webcolors==1.13
widgetsnbextension==4.0.9
Babel==2.12.1
Brlapi==0.8.5
Jinja2==3.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Mako==1.2.4
Markdown==3.5.1
MarkupSafe==2.1.3
Pillow==10.1.0
PyGObject==3.46.0
PyOpenGL==3.1.6
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySide2==5.15.11
PySide6==6.6.0
PyYAML==6.0.1
Pygments==2.16.1
Reflector==2023.6.28.0.36.1
Send2Trash==1.8.2
TBB==0.2
absl-py==2.0.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
argcomplete==2.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
autocommand==2.2.2
beautifulsoup4==4.12.2
bleach==6.1.0
btrfsutil==6.6.2
cachecontrol==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
coverage==6.5.0
cryptography==41.0.6
cssselect==1.2.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
envycontrol==3.3.1
executing==2.0.0
fastjsonschema==2.19.0
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.44.3
gast==0.3.3
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.10.0
html5lib==1.1
idna==3.4
importlib-metadata==5.0.0
inflect==7.0.0
injector==0.21.0
ipykernel==6.27.1
ipython==8.17.2
jaraco.context==4.3.0
jaraco.functools==3.9.0
jaraco.text==3.11.1
jedi==0.19.1
joblib==1.3.2
json5==0.9.14
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-client==8.6.0
jupyter-core==5.5.0
jupyter-events==0.9.0
jupyter-server==2.11.1
jupyterlab==4.0.9
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.2
keras==2.12.0
kiwisolver==1.4.5
lensfun==0.3.4
lit==16.0.6.dev0
lockfile==0.12.2
louis==3.27.0
lxml==4.9.2
markdown-it-py==3.0.0
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.3.1
more-itertools==10.1.0
msgpack==1.0.5
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.7
netsnmp-python==1.0a1
notebook-shim==0.2.3
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
ordered-set==4.1.0
overrides==7.4.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
peewee==3.16.3
pexpect==4.8.0
pip==23.3.1
platformdirs==4.0.0
ply==3.11
pooch==1.8.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pwquality==1.4.5
py3nvml==0.2.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.11.1
pycairo==1.25.1
pycparser==2.21
pydantic==2.5.2
pydantic-core==2.14.5
pydot==1.4.2
pyparsing==3.1.1
pyrsistent==0.19.3
pyserial==3.5
python-dateutil==2.8.2
python-json-logger==2.0.7
python-xlib==0.33
pytz==2023.3.post1
pyxdg==0.28
pyzmq==25.1.1
reactivex==4.0.4
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.10.6
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.7
scikit-learn==1.3.2
scipy==1.11.4
scour==0.38.2
setproctitle==1.3.3
setuptools==68.2.0
shiboken2==5.15.11
shiboken6==6.6.0
shiboken6-generator==6.6.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
tensorboard==2.15.0
tensorboard-data-server==0.8.0a0
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
termcolor==2.3.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
traitlets==5.14.0
trash-cli==0.23.2.13.2
trove-classifiers==2023.11.23
typing-extensions==4.8.0
uc-micro-py==1.0.2
urllib3==1.26.18
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
virtualenv==20.24.6
wcwidth==0.2.9
webencodings==0.5.1
websocket-client==1.6.4
werkzeug==2.3.8
wheel==0.40.0
wrapt==1.14.1
zipp==3.17.0
zstandard==0.22.0
